backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 环境变量管理
python-dotenv==1.0.0

# LangChain核心库
langchain-core==0.1.0
langchain-community==0.0.10
langchain-chroma==0.1.0
langchain-text-splitters==0.0.1

# 文档处理
pypdf==3.17.0
python-docx==1.1.0

# 其他依赖
pydantic==2.5.0

